286 Fixed weird blanked files
